+++ src/components/Textarea.jsx
@@ -4,6 +4,7 @@ export default function Textarea({
   value,
   type,
   name,
+  onBlur,
   onChange,
   placeholder,
   dataAttributes,
@@ -16,6 +17,7 @@ export default function Textarea({
       value={value}
       type={type}
       name={name}
+      onBlur={(e) => onBlur(e.currentTarget)}
       onChange={onChange}
       onKeyDown={(e) => e.key === 'Enter' && e.preventDefault()}
       placeholder={placeholder}